Use manually created variables for CI instead of CI_JOB_TOKEN
For protected branches, it seems now just CI_JOB_TOKEN is not enough.

https://gitlab.com/gitlab-org/gitlab-foss/-/issues/36898#note_38415655

According to this, the CI_JOB_TOKEN is based on whoever created the
job and creating a pipeline on a protected branch requires special
permissions. Somehow this still did not work for other people who
merged, even though they had access to the docs repo.